requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.8.1
    # via -r requirements.in
anyio==3.6.1
    # via starlette
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via click
fastapi==0.79.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
mako==1.2.1
    # via alembic
markupsafe==2.1.1
    # via mako
mysqlclient==2.1.1
    # via -r requirements.in
pydantic==1.9.1
    # via fastapi
pyjwt==2.4.0
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
six==1.16.0
    # via python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.39
    # via
    #   -r requirements.in
    #   alembic
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via pydantic
uvicorn==0.18.2
    # via -r requirements.in
